Move fork impl into serverProcess
