+++ extensions/typescript-language-features/src/utils/electron.ts
@@ -6,7 +6,6 @@
 import * as temp from './temp';
 import path = require('path');
 import fs = require('fs');
-import cp = require('child_process');
 import process = require('process');
 
 
@@ -38,35 +37,4 @@ export const getInstanceDir = (() => {
 
 export function getTempFile(prefix: string): string {
 	return path.join(getInstanceDir(), `${prefix}-${temp.makeRandomHexString(20)}.tmp`);
-}
-
-function generatePatchedEnv(env: any, modulePath: string): any {
-	const newEnv = Object.assign({}, env);
-
-	newEnv['ELECTRON_RUN_AS_NODE'] = '1';
-	newEnv['NODE_PATH'] = path.join(modulePath, '..', '..', '..');
-
-	// Ensure we always have a PATH set
-	newEnv['PATH'] = newEnv['PATH'] || process.env.PATH;
-
-	return newEnv;
-}
-
-export interface ForkOptions {
-	readonly cwd?: string;
-	readonly execArgv?: string[];
-}
-
-export function fork(
-	modulePath: string,
-	args: readonly string[],
-	options: ForkOptions,
-): cp.ChildProcess {
-	const newEnv = generatePatchedEnv(process.env, modulePath);
-	return cp.fork(modulePath, args, {
-		silent: true,
-		cwd: options.cwd,
-		env: newEnv,
-		execArgv: options.execArgv
-	});
 }
